requirements.txt
torch
torchvision
facenet-pytorch
wandb
lpips
scipy
pandas
numpy
Pillow
tqdm
torchmetrics[image]>=1.0
